requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
coverage==7.2.5
cycler==0.11.0
databases==0.7.0
databricks-cli==0.17.7
decorator==5.1.1
docker==6.1.1
entrypoints==0.4
exceptiongroup==1.1.1
fastapi==0.95.1
Flask==2.3.2
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.5.0
gcsfs==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
memoized-property==1.0.3
mlflow==2.3.1
multidict==6.0.4
mysql-connector-python==8.0.33
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pluggy==1.0.0
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
PyJWT==2.7.0
PyMySQL==1.0.3
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
querystring-parser==1.2.4
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.48
sqlparse==0.4.4
starlette==0.26.1
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
websocket-client==1.5.1
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
